add e2e test suite
The e2e tests are run using the `make test-e2e` command, and goes
through creating, updating, and deleting consul kv pairs.

For the update and delete tests, the git sha1 reference key has to be
deleted before running git2consul because currently the reference key is
the object compared to determine if a change has been made and if kv
pairs should be updated. When issue 31 is complete, the deletion should
no longer be required.